app_state: convert observe to work with json
Convert obsevation of the "desired" LightDB State path to directly monitor
the "reset_cumulative" value.

This is necessary because the LightDB Observe API call for the Golioth
Firmware SDK only allows JSON-type observation. A PR has been opened to
allow CBOR to be selected. This commit can be reverted when that option is
available.

Signed-off-by: Mike Szczys <mike@golioth.io>